graphql/requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-GraphQL==2.0.0
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
promise==2.2.1
python-dotenv==0.10.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.16.0
